refactor: FSD 패턴으로 컴포넌트 구조 개선
