+++ src/components/shared/layout/index.ts
@@ -1,0 +1,1 @@
+export { NavigationHeader } from './NavigationHeader';
