+++ src/components/shared/ui/NotFoundLogo.tsx
