+++ src/components/shared/ui/ProtectedRoute.tsx
