+++ src/components/shared/ui/index.ts
@@ -1,0 +1,4 @@
+export { default as FormField } from './FormField';
+export { default as InputWithError } from './InputWithError';
+export { ProtectedRoute } from './ProtectedRoute';
+export { NotFoundLogo } from './NotFoundLogo';
